+++ core/systemcontracts/upgrade.go
@@ -55,6 +55,8 @@ var (
 	lubanUpgrade = make(map[string]*Upgrade)
 
 	platoUpgrade = make(map[string]*Upgrade)
+
+	latestUpgrade = platoUpgrade
 )
 
 func init() {
@@ -695,6 +697,8 @@ func init() {
 			},
 		},
 	}
+
+	latestUpgrade[defaultNet] = latestUpgrade[rialtoNet]
 }
 
 func UpgradeBuildInSystemContract(config *params.ChainConfig, blockNumber *big.Int, statedb *state.StateDB) {