+++ eth/gasprice/gasprice_test.go
@@ -138,6 +138,7 @@ func newTestBackend(t *testing.T, londonBlock *big.Int, pending bool) *testBacke
 	config.GibbsBlock = nil
 	config.LubanBlock = nil
 	config.PlatoBlock = nil
+	config.HertzBlock = nil
 	engine := ethash.NewFaker()
 	db := rawdb.NewMemoryDatabase()
 	genesis, err := gspec.Commit(db)